\name{infotheo}
\alias{infotheo}

\title{Information Theory package}

\description{
  The package infotheo provide various estimators for computing information-theoretic measures from data
}
\author{
  Patrick E. Meyer
}
\seealso{\code{\link{entropy}}, \code{\link{condentropy}}, \code{\link{mutinformation}}, \code{\link{condinformation}}, \code{\link{multiinformation}}, \code{\link{interinformation}}, \code{\link{natstobits}}}

\references{
Meyer,  P. E.  (2008). Information-Theoretic Variable Selection and Network Inference from Microarray Data. PhD thesis of the Universite Libre de Bruxelles.
}     

\keyword{misc}
